requirements.txt
opencv-python
numpy
Pillow
pycryptodome
matplotlib
PyWavelets